requirements.txt
streamlit
ultralytics
numpy
opencv-python
Pillow